backend/pyproject.toml
[project]
name = "didacus-backend"
version = "0.1.0"
description = ""
authors = [
  { name = "Diego Dias", email = "diego.diasg@outlook.com" }
]
readme = "README.md"
# Ajustamos aqui incluindo <4.0.0
requires-python = ">=3.11,<4.0.0"

dependencies = [
  "django (>=5.2.4,<6.0.0)",
  "djangorestframework (>=3.16.0,<4.0.0)",
  "django-environ (>=0.12.0,<0.13.0)",
  "django-cors-headers (>=4.7.0,<5.0.0)",
  "pillow (>=11.3.0,<12.0.0)",
  "psycopg2-binary (>=2.9.10,<3.0.0)",
  "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
  "django-filter (>=25.1,<26.0)",
  "drf-spectacular (>=0.28.0,<0.29.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-django = "^4.11.1"
flake8 = "^7.3.0"

